# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Subito Slash Service
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths 
basePath: /subito/
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /slack:
    x-swagger-router-controller: slackController
    post:
      description: handles and pipes all Slack slash commands
      operationId: handleSlack
      parameters:
        - name: body
          in: body
          description: a generic Slack post body
          required: true
          schema:
            required:
              - text
              - token
            properties:
              token:
                type: string
                description: a Slack integration token to be validated against
              text:
                type: string
              team_id:
                type: string
              team_domain:
                type: string
              channel_id:
                type: string
              channel_name:
                type: string
              user_id:
                type: string
              user_name:
                type: string
              command:
                type: string
              response_url:
                type: string 
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SlackResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /day:
    # binds a127 app logic to a route
    x-swagger-router-controller: soupCalendarController
    get:
      description: get the soups for today
      # used as the method name of the controller
      operationId: getSoupsForToday
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SoupCalendarViewObject"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /day/{day}:
    # binds a127 app logic to a route
    x-swagger-router-controller: soupCalendarController
    get:
      description: get the soups for a day
      # used as the method name of the controller
      operationId: getSoupsForDay
      parameters:
        - name: day
          in: path
          description: "the day to get soups for :: example - tomorrow | 12-08-2017"
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SoupCalendarViewObject"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /all:
    x-swagger-router-controller: soupCalendarController
    get:
      description: Gets the fully stored soup calendar
      operationId: getAllSoups
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/SoupCalendarViewObject"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  SoupCalendarViewObject:
    type: object
    properties:
      text:
        type: string
      day:
        type: string
        format: date
      soups:
        type: array
        items:
          type: string
      soupsStr:
        type: string
  SlackResponse:
    type: object
    required:
      - text
    properties:
      text:
        type: string
  ErrorResponse:
    required:
      - text
    properties:
      text:
        type: string